refactor epic UTs to multiple files for readability
